// dream-travels/src/interfaces/trips.dto.interface.ts
export interface TripsDTO {
  readonly id: number;
  readonly itinerary: ItineraryDTO[];
  readonly description: string;
  readonly photo_url: string;
  readonly status: string;
  readonly title: string;
}

export interface ItineraryDTO {
  readonly day: string;
  readonly description: string;
  readonly location: string;
}